LPS-196133 Temporarily fix issues with d3 npm package
This fixes the failures in LPS-196133 by not deduplicating the d3 packages.
This way the consumers of those packages will have their own copy of them in
AMD format and won't be hit by LPS-196820 that only happens in ESM flavour.

When LPS-196820 is fixed we will deduplicate d3 packages again.