LPS-195526 Using only service builder models
We were mixing service builder models with ObjectEntries from the REST
layer for no real reason.

Now we are using only service builder models both for custom and system
object entries.

This means all of them have the same interface and the external reference
code can be retrieved in the same way.