LPS-194075 Use "Pages Hierarchy" as the default configuration of the Menu Display portlet if it is on a Page Template
Here is the behavior that it implements for Page Templates (for
non-PageTemplate pages, the behavior is the same as it was previously):
* If Private Pages are disabled:
** Show "Pages Hierarchy" and configure it for Public Pages Hierarchy as
   long as there is at least one Public Page on the Site (same behavior
   as non-PageTemplate pages)
* If Private Pages are enabled:
** Always show "Pages Hierarchy" if there is at least one page, because
   whether we use the Public Pages Hierarchy or Private Pages Hierarchy
   is based on whether the page inheriting it is public or private.
** Configure it for Public Pages Hierarchy if there is at least one
   public page. Otherwise, configure it for Private Pages Hierarchy.
** In a later commit in this PR, we will add an info alert to inform the
   user that the Menu Display will appear different on the page that
   inherits it if its privateLayout boolean is the opposite of what this
   Menu Display is configured for now.

$$breaking_change_keyword$$
What: The default configuration has changed such that if a Menu Display widget is placed on a Page Template, the default configuration for it will be "Pages Hierarchy" (unless there are no Pages at all on the Site). It will display as "Pages Hierarchy" even if Private Pages are enabled. If Private Pages are enabled, an alert will appear on the Page Template letting the user know that the Menu Display may appear different if the inherited page has the opposite `privateLayout` value to the menu currently being rendered.
By default, the Menu Display widget on the Page Template will display the Public Pages Hierarchy. However, if there are no Public Pages (but there are Private Pages) on the Site, then it will display the Private Pages Hierarchy instead. If there are no pages at all, then the "Pages Hierarchy" option will not appear, and the "Primary Navigation Menu" will be the default option (similar to how it already works on non-PageTemplate pages).
Why: Currently, the Menu Display widget's default configuration is misleading when placed on a Page Template. The default configuration will be either "Private Pages Hierarchy" if there are any Private Pages on the Site, or it will be "Primary Navigation Menu" if there aren't any Private Pages on the Site. It will never be "Public Pages Hierarchy" because Page Templates are considered as Private Pages.
If the configuration of the Menu Display widget is never modified, then it will be considered to be in the "default" configuration. This means that when a Public Page inherits the Page Template, the Menu Display will render the "Public Pages Hierarchy", even though on the Page Template it was rendering either the "Primary Navigation Menu" or "Private Pages Hierarchy" instead. This can pose a nasty surprise to the user if they were expecting it to render the "Primary Navigation Menu" as the Page Template seemed to show.
Furthermore, if there are Private Pages on the Site, the user is able to explicitly configure the Menu Display on the Page Template for "Private Pages Hierarchy" by changing the configuration away from "Private Pages Hierarchy" and then changing it back to "Private Pages Hierarchy". This means that if a Public Page inherits the Page Template, it will display the "Private Pages Hierarchy", which is something that should not be allowed on Public Pages.